+++ src/app/core/models/config.ts
@@ -1,0 +1,5 @@
+export interface ConfigModel {
+    id: string
+    email: string
+    login: string
+}